+++ src/main/java/com/aks/flightbookingservice/common/FlightConstants.java
@@ -1,0 +1,4 @@
+package com.aks.flightbookingservice.common;
+
+public class FlightConstants {
+}
