+++ src/main/java/com/aks/flightbookingservice/entities/BookingEntity.java
@@ -1,0 +1,35 @@
+package com.aks.flightbookingservice.entities;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.io.Serializable;
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.List;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Entity
+@Table(name = "BOOKING_RECORD")
+public class BookingEntity implements Serializable {
+    @Id
+    @GeneratedValue
+    private long bookingId;
+    private LocalDate bookingDate;
+    private String destination;
+    private String source;
+    private double fare;
+    private String flightDate;
+    private String flightNumber;
+    private LocalTime flightTime;
+    private String status;
+    @OneToMany(cascade = CascadeType.ALL)
+    @JoinTable(name = "bookingDetails", joinColumns = {@JoinColumn(name="bookingId")}, inverseJoinColumns = {@JoinColumn(name="passengerId")})
+    private List<PassengerEntity> passenger;
+}
