+++ src/main/java/com/aks/flightbookingservice/entities/BookingRecord.java
@@ -1,0 +1,4 @@
+package com.aks.flightbookingservice.entities;
+
+public class BookingRecord {
+}
