+++ src/main/java/com/aks/flightbookingservice/entities/CountryCode.java
@@ -1,0 +1,11 @@
+package com.aks.flightbookingservice.entities;
+
+public enum CountryCode {
+    DEL,
+
+    HYD,
+
+    GKP,
+
+    JPR
+}
