+++ src/main/java/com/aks/flightbookingservice/exception/handler/DuplicateFlightException.java
@@ -1,0 +1,10 @@
+package com.aks.flightbookingservice.exception.handler;
+
+/**
+ * Custom exception for DuplicateFlight.
+ */
+public class DuplicateFlightException extends RuntimeException {
+    public DuplicateFlightException(String message) {
+        super(message);
+    }
+}
