+++ src/main/java/com/aks/flightbookingservice/exception/handler/FlightNotFoundException.java
@@ -1,0 +1,10 @@
+package com.aks.flightbookingservice.exception.handler;
+
+/**
+ * Custom exception for FlightNotFound.
+ */
+public class FlightNotFoundException extends RuntimeException {
+    public FlightNotFoundException(String message) {
+        super(message);
+    }
+}
