+++ src/main/java/com/aks/flightbookingservice/mapper/FlightBookingMapper.java
@@ -1,0 +1,11 @@
+package com.aks.flightbookingservice.mapper;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
+
+/**
+ * Mapper implementation for flight service to convert entity to dto and vice versa.
+ */
+@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = "spring")
+public interface FlightBookingMapper {
+}
