+++ src/main/java/com/aks/flightbookingservice/service/UserService.java
@@ -1,0 +1,4 @@
+package com.aks.flightbookingservice.service;
+
+public interface UserService {
+}
