+++ src/main/java/com/aks/flightbookingservice/service/serviceImpl/UserServiceImpl.java
@@ -1,0 +1,4 @@
+package com.aks.flightbookingservice.service.serviceImpl;
+
+public class UserServiceImpl {
+}
