+++ src/main/java/com/aks/flightbookingservice/util/FareCalculationUtil.java
@@ -1,0 +1,7 @@
+package com.aks.flightbookingservice.util;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class FareCalculationUtil {
+}
