--- src/main/java/com/aks/flightdataservice/common/FlightConstants.java
@@ -1,4 +1,0 @@
-package com.aks.flightdataservice.common;
-
-public class FlightConstants {
-}
