--- src/main/java/com/aks/flightdataservice/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.aks.flightdataservice.controller;
-
-public class UserController {
-}
