--- src/main/java/com/aks/flightdataservice/entities/BookingEntity.java
@@ -1,4 +1,0 @@
-package com.aks.flightdataservice.entities;
-
-public class BookingEntity {
-}
