--- src/main/java/com/aks/flightdataservice/entities/BookingRecord.java
@@ -1,4 +1,0 @@
-package com.aks.flightdataservice.entities;
-
-public class BookingRecord {
-}
