--- src/main/java/com/aks/flightdataservice/entities/CountryCode.java
@@ -1,11 +1,0 @@
-package com.aks.flightdataservice.entities;
-
-public enum CountryCode {
-    DEL,
-
-    HYD,
-
-    GKP,
-
-    JPR
-}
