--- src/main/java/com/aks/flightdataservice/entities/FlightStatus.java
@@ -1,8 +1,0 @@
-package com.aks.flightdataservice.entities;
-
-public enum FlightStatus {
-    RUNNING,
-    CLOSED,
-    HOLD,
-    COMING_SOON
-}
