--- src/main/java/com/aks/flightdataservice/entities/PassengerEntity.java
@@ -1,4 +1,0 @@
-package com.aks.flightdataservice.entities;
-
-public class PassengerEntity {
-}
