--- src/main/java/com/aks/flightdataservice/exception/handler/DuplicateFlightException.java
@@ -1,10 +1,0 @@
-package com.aks.flightdataservice.exception.handler;
-
-/**
- * Custom exception for DuplicateFlight.
- */
-public class DuplicateFlightException extends RuntimeException {
-    public DuplicateFlightException(String message) {
-        super(message);
-    }
-}
