--- src/main/java/com/aks/flightdataservice/exception/handler/FlightNotFoundException.java
@@ -1,10 +1,0 @@
-package com.aks.flightdataservice.exception.handler;
-
-/**
- * Custom exception for FlightNotFound.
- */
-public class FlightNotFoundException extends RuntimeException {
-    public FlightNotFoundException(String message) {
-        super(message);
-    }
-}
