--- src/main/java/com/aks/flightdataservice/mapper/FlightBookingMapper.java
@@ -1,4 +1,0 @@
-package com.aks.flightdataservice.mapper;
-
-public interface FlightBookingMapper {
-}
