--- src/main/java/com/aks/flightdataservice/repository/FlightBookingRepository.java
@@ -1,4 +1,0 @@
-package com.aks.flightdataservice.repository;
-
-public interface FlightBookingRepository {
-}
