--- src/main/java/com/aks/flightdataservice/service/FlightService.java
@@ -1,14 +1,0 @@
-package com.aks.flightdataservice.service;
-
-import com.aks.model.FlightDataDto;
-import com.aks.model.FlightFullResponseDataDto;
-import com.aks.model.FlightListDataDto;
-
-/**
- * Flight data service.
- */
-public interface FlightService {
-    FlightFullResponseDataDto addNewAirLine(FlightDataDto flightDataDto);
-
-    FlightListDataDto getFlightByFlightNumber(String flightNumber);
-}
