--- src/main/java/com/aks/flightdataservice/service/UserService.java
@@ -1,4 +1,0 @@
-package com.aks.flightdataservice.service;
-
-public interface UserService {
-}
