--- src/main/java/com/aks/flightdataservice/service/serviceImpl/UserServiceImpl.java
@@ -1,4 +1,0 @@
-package com.aks.flightdataservice.service.serviceImpl;
-
-public class UserServiceImpl {
-}
