+++ src/main/resources/application.properties
@@ -5,6 +5,6 @@ spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.PostgreSQLDialect
 spring.datasource.url=jdbc:postgresql://localhost:5432/aks_airline
 spring.datasource.username=postgres
 spring.datasource.password=postgres
-server.port=8181
+server.port=8282
 springdoc.swagger-ui.url=/airline.yml
-server.servlet.context-path=/comflightds
+server.servlet.context-path=/combookingds
